fix: modal for auto network switching
